Uploading HTML, CSS
This will be the basic start of the web app. I will then keep on adding more things as we grow from this. One thing would be adding a JS file but write know I am focusing more on the aesthetics and looks before I get my hands dirty is JS.